==> CS_utils/Py_summary_code/11. A3.4 VMT Reduction-Rev.ipynb <==
import os
import pandas as pd
import numpy as np
import openmatrix as omx
import random
import yaml

from utility import *

import warnings
warnings.filterwarnings('ignore')

with open('config.yaml', 'r') as file:
    params = yaml.safe_load(file)
    
_join = os.path.join
_dir = os.path.dirname
_norm = os.path.normpath

highway_output_data = params['highway_output_data']
perf_measure_columns = params['final_columns']
periods = params['periods']
concept_id = params['concept_id']
summary_outputs = params['summary_dir']
filename_extension = params['filename_extension']

Change_in_Auto_VMT_Assignment_Based = -583852
Change_in_Auto_VMT_Mode_Choice_Based_Proposed_Methodology = -765659

df_vmt = pd.DataFrame({'Value':[Change_in_Auto_VMT_Assignment_Based,Change_in_Auto_VMT_Mode_Choice_Based_Proposed_Methodology]})

df_vmt

df_vmt['Concept_ID'] = concept_id
df_vmt['Metric_ID'] = 'A3.4'
df_vmt['Metric_name'] = ''
df_vmt['Submetric'] = 'A3.4.1'
df_vmt.at[0, 'Description'] = 'Change in Auto VMT - Assignment_Based'
df_vmt.at[1, 'Description'] = 'Change in Auto VMT - Mode Choice-Based (Proposed Methodology)'
df_vmt['Population'] = 'Whole Population'
df_vmt['Geography'] = 'Regional'
df_vmt['Zone_ID'] = ''
df_vmt['Origin_zone'] = ''
df_vmt['Dest_zone'] = ''
df_vmt['Purpose'] = ''
df_vmt['Mode'] = ''
df_vmt['Income'] = ''
df_vmt['Period'] = 'All'
df_vmt['Units'] = 'vehicle miles travelled'
df_vmt['Total_Increment'] = ''

summary_outputs

df_vmt = df_vmt[perf_measure_columns]
df_vmt.to_csv(_join(summary_outputs, 'A3.4.1' + '_weekday_VMT_' + concept_id + '_region' + filename_extension + '.csv'), index=False)

df_vmt['Value'] = df_vmt['Value'] * params['annual_auto_factor']
df_vmt['Submetric'] = 'A3.4.2'

df_vmt.to_csv(_join(summary_outputs, 'A3.4.2' + '_annual_VMT_' + concept_id + '_region' + filename_extension + '.csv'), index=False)
