==> CS_utils/Py_summary_code/97. CompareLogSums.ipynb <==
import os
import pandas as pd
import numpy as np
import openmatrix as omx
import random
import yaml
from pathlib import Path
from utility import *
import pickle

import warnings
warnings.filterwarnings('ignore')

import pandas as pd
pd.options.display.float_format = '{:,.3f}'.format
pd.set_option("display.max_rows", None)
pd.set_option("display.max_columns", None)

with open('config.yaml', 'r') as file:
    params = yaml.safe_load(file)
    
_join = os.path.join

# paths
ctramp_dir = params['ctramp_dir']
model_outputs_dir = params['model_dir']
summary_dir = params['summary_dir']
concept_id = params['concept_id']
preprocess_dir = _join(ctramp_dir, '_pre_process_files')

Path(summary_dir).mkdir(parents=True, exist_ok=True)
Path(preprocess_dir).mkdir(parents=True, exist_ok=True)

purpose = ['Work', 'University', 'School', 'Escort', 'Shopping', 'EatOut', 
           'OthMaint', 'Social', 'OthDiscr', 'WorkBased']

time_period = {1:'EA',2:'AM',3:'MD',4:'PM',5:'EV'} #1 for EA, 2 for AM, 3 for MD, 4 for PM and 5 for EV

%%time


roh_dict = {}

for tripPeriod in time_period:
    
    tp_name = time_period[tripPeriod]
    print(f"reading baseline logsums")
    with open(_join(f'logsums_'+ tp_name +'_BL.pkl'), 'rb') as file:
        ls_bl = pickle.load(file)

    print(f"reading concept logsums")
    with open(_join(preprocess_dir, f'logsums_'+ tp_name +'.pkl'), 'rb') as file:
        ls_cpt = pickle.load(file)

    print(f"reading baseline trips")
    with open(_join(f'trips_'+ tp_name +'_BL.pkl'), 'rb') as file:
        trips_bl = pickle.load(file)

    print(f"reading concept trips")
    with open(_join(preprocess_dir, f'trips_'+ tp_name +'.pkl'), 'rb') as file:
        trips_cpt = pickle.load(file)

    for purp in purpose:
        purp_bl = ls_bl[purp]
        purp_cpt = ls_cpt[purp]
        trp_bl = trips_bl[purp]
        trp_cpt = trips_cpt[purp]

        auto_proportion_diff = purp_cpt['auto_proportion'] - purp_bl['auto_proportion']
        trn_proportion_diff = purp_cpt['trn_proportion'] - purp_bl['trn_proportion']
        nm_proportion_diff = purp_cpt['nm_proportion'] - purp_bl['nm_proportion']
        rh_proportion_diff = purp_cpt['rh_proportion'] - purp_bl['rh_proportion']

        auto_trips = trp_cpt['auto_trips'] - trp_bl['auto_trips']
        trn_trips = trp_cpt['trn_trips'] - trp_bl['trn_trips']
        nm_trips = trp_cpt['nm_trips'] - trp_bl['nm_trips']
        rh_trips = trp_cpt['rh_trips'] - trp_bl['rh_trips']

        auto_trips[auto_trips<0] = 0
        trn_trips[trn_trips<0] = 0
        nm_trips[nm_trips<0] = 0
        rh_trips[rh_trips<0] = 0
        
        print(np.min(auto_trips), np.max(auto_trips))
        print(np.min(trn_trips), np.max(trn_trips))
        print(np.min(nm_trips), np.max(nm_trips))
        print(np.min(rh_trips), np.max(rh_trips))

        auto_benefits = (trp_bl['auto_trips'] + auto_trips) * auto_proportion_diff * 0.5 /0.025
        trn_benefits = (trp_bl['trn_trips'] + trn_trips) * trn_proportion_diff * 0.5 / 0.025
        nm_benefits = (trp_bl['nm_trips'] + nm_trips) * nm_proportion_diff * 0.5 / 0.025
        rh_benefits = (trp_bl['rh_trips'] + rh_trips) * rh_proportion_diff * 0.5 / 0.025

        data_dict = {'auto_benefits' : auto_benefits,
                     'trn_benefits' : trn_benefits,
                     'nm_benefits' : nm_benefits,
                     'rh_benefits' : rh_benefits}

        roh_dict[purp] = data_dict
        
        print(purp, ",", tp_name ,",", np.sum(auto_benefits), ",", np.sum(trn_benefits), np.sum(nm_benefits), np.sum(rh_benefits))
    
    file_path = os.path.join(preprocess_dir, f"benefits_{tp_name}.pkl")
    with open(file_path, "wb") as pickle_file:
        pickle.dump(roh_dict, pickle_file, protocol=pickle.HIGHEST_PROTOCOL)
          
    print(f"Data has been written to {file_path} for {purp} and {tripPeriod}.") 
             

# non intrazonal 

for tripPeriod in time_period:, 
    
    tp_name = time_period[tripPeriod]
    with open(_join(preprocess_dir, f'benefits_'+ tp_name +'.pkl'), 'rb') as file:
        benefits = pickle.load(file)
        
    for purp in purpose:
        purp_benefit = benefits[purp]
        auto_benefits = purp_benefit['auto_benefits']
        trn_benefits = purp_benefit['trn_benefits']
        nm_benefits = purp_benefit['nm_benefits']
        rh_benefits = purp_benefit['rh_benefits']
        
        np.fill_diagonal(auto_benefits, 0)
        np.fill_diagonal(trn_benefits, 0)
        np.fill_diagonal(nm_benefits, 0)
        np.fill_diagonal(rh_benefits, 0)
        
        print(purp, ",", tp_name ,",", np.sum(auto_benefits), ",", np.sum(trn_benefits), np.sum(nm_benefits), np.sum(rh_benefits))

%%time
#QAQC
for tripPeriod in time_period:
    
    tp_name = time_period[tripPeriod]
    with open(_join(preprocess_dir, f'benefits_'+ tp_name +'.pkl'), 'rb') as file:
        benefits = pickle.load(file)
        
    for purp in purpose:
        purp_benefit = benefits[purp]
        auto_benefits = purp_benefit['auto_benefits']
        trn_benefits = purp_benefit['trn_benefits']
        nm_benefits = purp_benefit['nm_benefits']
        rh_benefits = purp_benefit['rh_benefits']
        
        np.fill_diagonal(auto_benefits, 0)
        np.fill_diagonal(trn_benefits, 0)
        np.fill_diagonal(nm_benefits, 0)
        np.fill_diagonal(rh_benefits, 0)
        
        print(purp, ",", tp_name ,",", np.max(auto_benefits), np.min(auto_benefits), np.mean(auto_benefits))
        print(purp, ",", tp_name ,",", np.max(trn_benefits), np.min(trn_benefits), np.mean(trn_benefits))
        print(purp, ",", tp_name ,",", np.max(nm_benefits), np.min(nm_benefits), np.mean(nm_benefits))
        print(purp, ",", tp_name ,",", np.max(rh_benefits), np.min(rh_benefits), np.mean(rh_benefits))

for tripPeriod in time_period:
    
    tp_name = time_period[tripPeriod]
    print(f"reading baseline logsums")
    with open(_join(f'logsums_'+ tp_name +'_BL.pkl'), 'rb') as file:
        ls_bl = pickle.load(file)

    print(f"reading concept logsums")
    with open(_join(preprocess_dir, f'logsums_'+ tp_name +'.pkl'), 'rb') as file:
        ls_cpt = pickle.load(file)

    print(f"reading baseline trips")
    with open(_join(f'trips_'+ tp_name +'_BL.pkl'), 'rb') as file:
        trips_bl = pickle.load(file)

    print(f"reading concept trips")
    with open(_join(preprocess_dir, f'trips_'+ tp_name +'.pkl'), 'rb') as file:
        trips_cpt = pickle.load(file)

    for purp in purpose:
        purp_bl = ls_bl[purp]
        purp_cpt = ls_cpt[purp]
        trp_bl = trips_bl[purp]
        trp_cpt = trips_cpt[purp]
        
        for keys in purp_bl:
