==> CS_utils/Py_summary_code/98_Performance_Metrics_Summary.ipynb <==
import pandas as pd
import numpy as np
import os
print("Import done")

import yaml

with open('config.yaml', 'r') as file:
    params = yaml.safe_load(file)
    
_join = os.path.join    
summary_outputs = params['summary_dir']
ctramp_dir = params['ctramp_dir']
iteration = params['iteration']

## Change working directory to the location where files are located
#os.chdir(summary_outputs)
#print(os.getcwd())

## Define population year and calculate the total population
year = 2050
total_population = len(pd.read_csv(_join(ctramp_dir, 'main\\personData_' + str(iteration) + '.csv')))

print(year, total_population)

## Define the metric names and their associated fitlers to the  'dict_of_metrics' dictionary below

dict_of_metrics = { 'A1.1.1':{'Period':'All','Income':'All'}, #Region
                    'A1.1.2':{'Period':'All','Income':'All'}, #Transbay Region
                    #'A1.1.3':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #County (No zone ID thus results in multiple rows)
                    #'A1.1.4':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #RDM Zone
                    #'A1.1.5':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #Super District
                    'A1.1.6':{'Period':'All','Income':'All'}, #Priority Population
                    #'A1.1.7':{'Period':'All','Income':'All'},#RDM Zone Priority Populaiton #Refine and use later

                    'A1.2.1':{'Period':'All','Income':'All'}, #Region
                    'A1.2.2':{'Period':'All','Income':'All'}, #Transbay Region
                    #'A1.2.3':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #County (No zone ID thus results in multiple rows)
                    #'A1.2.4':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #RDM Zone
                    #'A1.2.5':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #Super District
                    'A1.2.6':{'Period':'All','Income':'All'}, #Priority Population
                    #'A1.2.7':{'Period':'All','Income':'All'},#RDM Zone Priority Populaiton #Refine and use later

                    'A1.3.1':{'Period':'All','Income':'All'}, #Region
                    'A1.3.2':{'Period':'All','Income':'All'}, #Transbay Region
                    #'A1.3.3':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #County (No zone ID thus results in multiple rows)
                    #'A1.3.4':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #RDM Zone
                    #'A1.3.5':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #Super District
                    'A1.3.6':{'Period':'All','Income':'All'}, #Priority Population
                    #'A1.3.7':{'Period':'All','Income':'All'},#RDM Zone Priority Populaiton #Refine and use later

                    'A1.4.1':{'Period':'All','Income':'All'}, #Region
                    'A1.4.2':{'Period':'All','Income':'All'}, #Transbay Region
                    #'A1.4.3':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #County (No zone ID thus results in multiple rows)
                    #'A1.4.4':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #RDM Zone
                    #'A1.4.5':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #Super District
                    'A1.4.6':{'Period':'All','Income':'All'}, #Priority Population
                    #'A1.4.7':{'Period':'All','Income':'All'},#RDM Zone Priority Populaiton #Refine and use later

                    'A1.5.1':{'Period':'All','Income':'All'}, #Region
                    'A1.5.2':{'Period':'All','Income':'All'}, #Transbay Region
                    #'A1.5.3':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #County (No zone ID thus results in multiple rows)
                    #'A1.5.4':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #RDM Zone
                    #'A1.5.5':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #Super District
                    'A1.5.6':{'Period':'All','Income':'All'}, #Priority Population
                    #'A1.5.7':{'Period':'All','Income':'All'},#RDM Zone Priority Populaiton #Refine and use later
 
                    'A1.6.1':{'Period':'All','Income':'All'}, #Region
                    'A1.6.2':{'Period':'All','Income':'All'}, #Transbay Region
                    #'A1.6.3':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #County (No zone ID thus results in multiple rows)
                    #'A1.6.4':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #RDM Zone
                    #'A1.6.5':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #Super District
                    'A1.6.6':{'Period':'All','Income':'All'}, #Priority Population
                    #'A1.6.7':{'Period':'All','Income':'All'},#RDM Zone Priority Populaiton #Refine and use later

                    'A1.7.1':{'Period':'All','Income':'All'}, #Region
                    'A1.7.2':{'Period':'All','Income':'All'}, #Transbay Region
                    #'A1.7.3':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #County (No zone ID thus results in multiple rows)
                    #'A1.7.4':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #RDM Zone
                    #'A1.7.5':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #Super District
                    'A1.7.6':{'Period':'All','Income':'All'}, #Priority Population
                    #'A1.7.7':{'Period':'All','Income':'All'},#RDM Zone Priority Populaiton #Refine and use later

                    'A1.8.1':{'Period':'All','Income':'All'}, #Region
                    'A1.8.2':{'Period':'All','Income':'All'}, #Transbay Region
                    #'A1.8.3':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #County (No zone ID thus results in multiple rows)
                    #'A1.8.4':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #RDM Zone
                    #'A1.8.5':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #Super District
                    'A1.8.6':{'Period':'All','Income':'All'}, #Priority Population
                    #'A1.8.7':{'Period':'All','Income':'All'},#RDM Zone Priority Populaiton #Refine and use later

                    'A1.10.1':{'Period':'All','Income':'All'}, #Weekday regional crowding 
                    'A1.10.2':{'Period':'All','Income':'All'}, #Weekday regional crowding Priority Population
                    #'A1.10.3':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #Weekday regional crowding RDM zone
                    #'A1.10.4':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #Weekday regional crowding RDM zone Priority Population

                   'A3.1.1':{'Period':'All','Income':'All','Purpose':'All','Mode':'All'}, #Weekday Region
                   'A3.1.2':{'Period':'All','Income':'All','Purpose':'All','Mode':'All'}, #Weekday Transbay Region
                   #'A3.1.3':{'Period':'All','Income':'All','Purpose':'All','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Weekday County (No zone ID thus results in multiple rows)
                   #'A3.1.4':{'Period':'All','Income':'All','Purpose':'All','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Weekday RDM Zone
                   #'A3.1.5':{'Period':'All','Income':'All','Purpose':'All','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Weekday Super District
                   'A3.1.6':{'Period':'All','Income':'All','Purpose':'All','Mode':'All'}, #Weekday Priority Population
                   #'A3.1.7':{'Period':'All','Income':'All','Purpose':'All','Mode':'All'}, #Weekday RDM Zone Priority Population #Refine and use later
                   'A3.1.8':{'Period':'All','Income':'All','Purpose':'All','Mode':'All'}, #Annual Values for Region ####ERROR A3.1.8 files has A3.1.1
                   'A3.1.9':{'Period':'All','Income':'All','Purpose':'All','Mode':'All'}, #Annual Transbay Region
                   #'A3.1.10':{'Period':'All','Income':'All','Purpose':'All','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Annual County (No zone ID thus results in multiple rows)
                   #'A3.1.11':{'Period':'All','Income':'All','Purpose':'All','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Annual RDM Zone
                   #'A3.1.12':{'Period':'All','Income':'All','Purpose':'All','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Annual Super District
                   'A3.1.13':{'Period':'All','Income':'All','Purpose':'All','Mode':'All'}, #Annual Priority Population
                   #'A3.1.14':{'Period':'All','Income':'All','Purpose':'All','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Annual RDM Zone Priority Population
                   #'A3.1.15':{'Period':'All','Income':'All','Purpose':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Weekday trips in RDM zones by all modes

                    
                    'A3.2.1':{'Period':'AM'}, #Weekday Boardings
                   'A3.2.2':{'Period':'AM'}, #Weekday Alightings
                   #'A3.2.3':{'Period':'AM'}, #Weekday transit segment flows
                   'A3.2.4':{'Period':'AM'}, #Annual boardings
                   'A3.2.5':{'Period':'AM'}, #Annual alightings
             

                    'A3.3.1':{'Period':'All','Income':'All','Purpose':'All'}, #Region
                    'A3.3.2':{'Period':'All','Income':'All','Purpose':'All'}, #Transbay Region
                    #'A3.3.3':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #County
                    #'A3.3.4':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #RDM Zone
                    #'A3.3.5':{'Period':'All','Income':'All', 'Origin_zone':'All', 'Dest_zone':'All'}, #Super District
                    'A3.3.6':{'Period':'All','Income':'All','Purpose':'All'}, #Priority Population
                    #'A3.3.7':{'Period':'All','Income':'All'},#RDM Zone Priority Populaiton #Refine and use later
                    #'A3.3.8':{'Income':'All'},#RDM Zone Priority Populaiton ####ERROR Its by percentage segmented by mode

                    'B1.1.1':{'Period':'am','Description':'Total employment accessible within actual 60 minutes'}, #Region 
                    'B1.1.2':{'Period':'am','Description':'Total employment accessible within actual 60 minutes'}, #County 
                    #'B1.1.3':{'Period':'am','Description':'Total employment accessible within actual 60 minutes'}, #RDM Zone 
                    #'B1.1.4':{'Period':'am','Description':'Total employment accessible within actual 60 minutes'}, #Super District 
                    'B1.1.5':{'Period':'am','Description':'Total employment accessible within actual 60 minutes'}, #Priority Population


                   'B1.2.1':{'Period':'All','Income':'All','Purpose':'work','Mode':'All'}, #Weekday Region
                   'B1.2.2':{'Period':'All','Income':'All','Purpose':'work','Mode':'All'}, #Weekday transbay Region
                   #'B1.2.3':{'Period':'All','Income':'All','Purpose':'work','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Weekday County
                   #'B1.2.4':{'Period':'All','Income':'All','Purpose':'work','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Weekday RDM Zone
                   #'B1.2.5':{'Period':'All','Income':'All','Purpose':'work','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Weekday Super District
                   'B1.2.6':{'Period':'All','Income':'All','Purpose':'work','Mode':'All'}, #Weekday Priority Population
                   #'B1.2.7':{'Period':'All','Income':'All','Purpose':'work','Mode':'All'}, #Weekday RDM zone priority population
                   'B1.2.8':{'Period':'All','Income':'All','Purpose':'work','Mode':'All'}, # Annual values for region
                   'B1.2.9':{'Period':'All','Income':'All','Purpose':'work','Mode':'All'}, #Annual transbay region
                   #'B1.2.10':{'Period':'All','Income':'All','Purpose':'work','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Annual county
                   #'B1.2.11':{'Period':'All','Income':'All','Purpose':'work','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Annual RDM zone
                   #'B1.2.12':{'Period':'All','Income':'All','Purpose':'work','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Annual Super district
                   'B1.2.13':{'Period':'All','Income':'All','Purpose':'work','Mode':'All'}, #Annual Priority population
                   #'B1.2.14':{'Period':'All','Income':'All','Purpose':'work','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'},#Annual RDM zone priority population

                   'B1.4.1':{'Period':'All','Income':'All','Purpose':'non-work','Mode':'All'}, #Weekday Region
                   'B1.4.2':{'Period':'All','Income':'All','Purpose':'non-work','Mode':'All'}, #Weekday transbay Region
                   #'B1.4.3':{'Period':'All','Income':'All','Purpose':'non-work','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Weekday County
                   #'B1.4.4':{'Period':'All','Income':'All','Purpose':'non-work','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Weekday RDM Zone
                   #'B1.4.5':{'Period':'All','Income':'All','Purpose':'non-work','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Weekday Super District
                   'B1.4.6':{'Period':'All','Income':'All','Purpose':'non-work','Mode':'All'}, #Weekday Priority Population
                   #'B1.4.7':{'Period':'All','Income':'All','Purpose':'non-work','Mode':'All'}, #Weekday RDM zone priority population
                   'B1.4.8':{'Period':'All','Income':'All','Purpose':'non-work','Mode':'All'}, # Annual values for region
                   'B1.4.9':{'Period':'All','Income':'All','Purpose':'non-work','Mode':'All'}, #Annual transbay region
                   #'B1.4.10':{'Period':'All','Income':'All','Purpose':'non-work','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Annual county
                   #'B1.4.11':{'Period':'All','Income':'All','Purpose':'non-work','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Annual RDM zone
                   #'B1.4.12':{'Period':'All','Income':'All','Purpose':'non-work','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Annual Super district
                   'B1.4.13':{'Period':'All','Income':'All','Purpose':'non-work','Mode':'All'}, #Annual Priority population
                   #'B1.4.14':{'Period':'All','Income':'All','Purpose':'non-work','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'},#Annual RDM zone priority population 
                   
                   'B2.3':{'Description':'Total walk distance by person per day'},#'Period':'am',
                   'B2.4':{'Description':'Count of people walking for more than 30 minutes per day'}, #'Period':'am'
                   
                   'C1.2.1':{'Period':'am','Description':'Business access to potential employees within actual 60 minutes'}, #Region
                   'C1.2.2':{'Period':'am','Description':'Business access to potential employees within actual 60 minutes'}, #County 
                   #'C1.2.3':{'Period':'am','Description':'Business access to potential employees within actual 60 minutes'}, #RDM Zone
                   #'C1.2.4':{'Period':'am','Description':'Business access to potential employees within actual 60 minutes'}, # Super District


                   'C2.2.1':{'Period':'All','Income':'All','Purpose':'business','Mode':'All'}, #Weekday Region
                   'C2.2.2':{'Period':'All','Income':'All','Purpose':'business','Mode':'All'}, #Weekday transbay Region
                   #'C2.2.3':{'Period':'All','Income':'All','Purpose':'business','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Weekday County
                   #'C2.2.4':{'Period':'All','Income':'All','Purpose':'business','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Weekday RDM Zone
                   #'C2.2.5':{'Period':'All','Income':'All','Purpose':'business','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Weekday Super District
                   'C2.2.6':{'Period':'All','Income':'All','Purpose':'business','Mode':'All'}, #Weekday Priority Population
                   #'C2.2.7':{'Period':'All','Income':'All','Purpose':'business','Mode':'All'}, #Weekday RDM zone priority population
                   'C2.2.8':{'Period':'All','Income':'All','Purpose':'business','Mode':'All'}, # Annual values for region
                   'C2.2.9':{'Period':'All','Income':'All','Purpose':'business','Mode':'All'}, #Annual transbay region
                   #'C2.2.10':{'Period':'All','Income':'All','Purpose':'business','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Annual county
                   #'C2.2.11':{'Period':'All','Income':'All','Purpose':'business','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Annual RDM zone
                   #'C2.2.12':{'Period':'All','Income':'All','Purpose':'business','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Annual Super district
                   'C2.2.13':{'Period':'All','Income':'All','Mode':'All'}, #Annual Priority population
                   #'C2.2.14':{'Period':'All','Income':'All','Purpose':'business','Mode':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Annual RDM zone priority population 

                   'E1.5.1':{'Geography':'Region'},
                   'E1.5.2':{'Geography':'Region'},


                   'F1.1.1':{'Period':'All'}, #Region
                   'F1.1.2':{'Period':'All'}, #Transbay Region
                   #'F1.1.3':{'Period':'All','Origin_zone':'All', 'Dest_zone':'All'}, #County
                   #'F1.1.4':{'Period':'All','Origin_zone':'All', 'Dest_zone':'All'}, #RDM Zone
                   #'F1.1.5':{'Period':'All','Origin_zone':'All', 'Dest_zone':'All'}, #Super district
                   'F1.1.6':{'Period':'All'}, #Annual values for region
                   'F1.1.7':{'Period':'All'}, #Annual Transbay Region
                   'F1.1.8':{'Period':'All'}, #Annual County
                   #'F1.1.9':{'Period':'All'}, #Annual RDM zone
                   #'F1.1.10':{'Period':'All','Origin_zone':'All', 'Dest_zone':'All'} #Annual Super District
                   
                   'G1.1.1':{'Period':'All'}, #weekday linked rail trips
                   'G1.1.2':{'Period':'All'}, #Annual linked rail trips

                   }

## Add the metric names to 'max_agg_of_metrics' list if they need to be summzrized by max value
max_agg_of_metrics = ['B1.1.1']

## Add the metric names to 'divide_by_pop' list if they need to be divided by population
divide_by_pop = ['B2.3']

## Add the metric names to 'summarize_by' list if they will be summarized by each category such as 'Zone_ID', 'Mode', etc.
summarize_by = {'C1.2.2':'Zone_ID',
                 'A3.3.1':'Mode',
                  'A3.3.2':'Mode',
                  'A3.3.6':'Mode',
                  }

metric_name = []
metric_val = []
total_values = []
special_case = []
category_name = []
category_val = []
actual_perceived = []
all_filter_conditions = []
description = []

files = os.listdir(summary_outputs)
for new_file in files:
    split_file_name = new_file.split('_') #To identiy the metric

    for metric in list(dict_of_metrics.keys()): #Gathers the list of metrics defined in above dictionary
        if str(metric) in split_file_name: #Read the file if listed metric is found
            df = pd.read_csv(_join(summary_outputs,new_file))

            # Check if the file name mentions as actual or perceived
            if 'actual' in new_file:
                actual_perceived_val = 'actual'
            elif 'perceived' in new_file:
                actual_perceived_val = 'perceived'
            else:
                actual_perceived_val = ''

            # Extract the performance metric through sequential filtering of the data
            filter_conditions_this_metric = []
            for key in list(dict_of_metrics[metric].keys()): #Gathers the defined filters for each metric
                condition = key
                filter_key = dict_of_metrics[metric][key]
                df = df[df[condition]==filter_key]
                filter_conditions_this_metric.append((condition, filter_key))
            output_values = df['Value'].values

            if df.shape[0] > 0:
                description_val = df.Description.values[0]

            # Condition for metric if it need not be summarized by categories such as 'Zone_ID', 'Mode'
            if metric not in list(summarize_by.keys()):
                if len(output_values) == 0:
                    break
                elif len(output_values)>1:
                    if metric in max_agg_of_metrics:
                        final_output = max(output_values)
                    else:
                        final_output = sum(output_values)
                elif metric in divide_by_pop:
                    final_output = output_values[0]/total_population
                else:
                    final_output = output_values[0]

                metric_val.append('{0:.2f}'.format(final_output))
                metric_name.append(metric)
                total_values.append(len(output_values))
                category_name.append('')
                category_val.append('')
                actual_perceived.append(actual_perceived_val)
                all_filter_conditions.append(filter_conditions_this_metric)
                description.append(description_val)
            
            # Condition for metric if it must summarized by categories such as 'Zone_ID', 'Mode' as listed in 'summarize_by' dictionary
            if metric in list(summarize_by.keys()):

                for index, row in df.iterrows():
                    metric_val.append('{0:.2f}'.format(row['Value']))
                    metric_name.append(metric)
                    total_values.append(1)
                    category_name.append(summarize_by[metric])
                    category_val.append(row[summarize_by[metric]])
                    actual_perceived.append(actual_perceived_val)
                    all_filter_conditions.append(filter_conditions_this_metric)
                    description.append(description_val)
                                 


print(len(metric_name))
print(len(metric_val))
print(len(category_name))
print(len(total_values))
print(len(category_val))
print(len(actual_perceived))
print(len(all_filter_conditions))
print(len(description))



##Output the summaries
store_df = pd.DataFrame({'Metric_Name':metric_name,
                          'Metric_Value': metric_val,
                          'Description':description,
                          'total_values':total_values,
                          'category_name':category_name,
                          'category_val':category_val,
                          'actual_perceived':actual_perceived,
                          'filter_conditions':all_filter_conditions})
store_df.to_csv(r'combined_perf_measures_'+params['concept_id']+'.csv', index=False)
