==> CS_utils/Py_summary_code/99. TripSummaries.ipynb <==
#### Set the Constants

out_file_name = 'Trip Summaries2.xlsx'
# out_dir = r'C:\Users\msaeidirazavi\OneDrive - Cambridge Systematics\MyFiles\210071 Link21\Trip Summary Scripts\Outputs'

quick = 0   # 0 or 1. Choose zero for full run,
            # else results will be just for a sample from the original input dataframe.

sample_size = 10001   # Choose a number greater than 10,000 or else the script may throw an error. 
c1, c2 = 'Measure_Description', 'Value'  # Column names in the output excel file.

#### Import Python Libraries

import os
import pandas as pd
import yaml
from collections import Counter

#### Set Pandas Options

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

#### Determine Input and Output Paths

with open('config.yaml', 'r') as file:
    params = yaml.safe_load(file)
    
_join = os.path.join

ctramp_dir = params['ctramp_dir']
in_trip_roster_path = _join(ctramp_dir, "_pre_process_files\\trip_roster.parquet")
in_link_vols = params['highway_output_data']   

summary_outputs = params['summary_dir']
out_path = _join(summary_outputs, out_file_name)

#### Create the Summary

df_out = pd.DataFrame({c1:[], c2:[]})
df_out

in_link_vols

df_original = pd.read_parquet(in_trip_roster_path)
df_links = pd.read_csv(in_link_vols)

df_sample = df_original.sample(n=sample_size)
df_sample.head()

df = df_sample if quick else df_original 

total_trips = df['trips'].sum()
print(f'Total number of trips: {total_trips:,}')
new = pd.DataFrame({c1:['Trips_Total'], c2:[total_trips]})
df_out = pd.concat([df_out, new])
df_out

modes = list(df['Mode'].unique()) 
modes = sorted(modes)
modes

modes = [
    'Walk',
    'Bike',
    'WALK_TRANSIT',
    'KNR_TRANSIT',
    'PNR_TRANSIT',    
    'Auto_3+Person',
    'Auto_2Person',
    'Auto_SOV',
    'RIDE_HAIL'
]

modes_auto = ['Auto_SOV', 'Auto_2Person', 'Auto_3+Person']
modes_transit = ['WALK_TRANSIT', 'PNR_TRANSIT', 'KNR_TRANSIT']

#counts = Counter(df["Mode"])
counts = df.groupby(["Mode"])['trips'].sum().reset_index()
dic_counts = dict(zip(counts['Mode'], counts['trips']))
dic_counts

transit_trips = 0
for m in modes_transit:
    transit_trips += dic_counts[m]
print(f'Total number of transit trips: {transit_trips:,}')
new = pd.DataFrame({c1:['Trips_Transit'], c2:[transit_trips]})
df_out = pd.concat([df_out, new])
df_out

tb_trips = df.loc[(df['transbay_od']==1)].copy()
tb_trips = tb_trips['trips'].sum()
print(f'Total number of transbay trips: {tb_trips:,}')

tb_transit_trips = df.loc[(df['transbay_od']==1)&(df['Mode'].str.contains('TRANSIT'))].copy()
tb_transit_trips = tb_transit_trips['trips'].sum()
print(f'Transbay transit trips: {tb_transit_trips:,}')

tb_auto_trips = df.loc[(df['transbay_od']==1)&(df['Mode'].str.contains('Auto'))].copy()                                                                                              
tb_auto_trips = tb_auto_trips['trips'].sum()
print(f'Transbay auto trips: {tb_auto_trips:,}')

new = pd.DataFrame({c1:['Trips_Transbay_Total', 'Trips_Transbay_Transit', 'Trips_Transbay_Auto'],
                    c2:[tb_trips, tb_transit_trips, tb_auto_trips]})
df_out = pd.concat([df_out, new])
df_out

df_shares = pd.DataFrame(list(dic_counts.items()), columns=['Mode', 'TripCount'])
df_shares

df_shares['ModeShare'] = df_shares['TripCount']/total_trips
df_shares

for m in modes:
    for i, row in df_shares.iterrows():
        if row['Mode'] == m:
            desc = f'ModePercentShare_{m}'
            val = round(100 * row['ModeShare'], 5)
            new = pd.DataFrame({c1:[desc], c2:[val]})
            df_out = pd.concat([df_out, new])
            break
df_out

tods = params['periods']

dic_tod = {}
for tod in tods:
    df_temp = df.loc[df['Period']==tod]
    val = df_temp['trips'].sum()
    dic_tod[tod] = val
    desc = f'TOD_{tod}_Trips'
    new = pd.DataFrame({c1:[desc], c2:[val]})
    df_out = pd.concat([df_out, new])

for tod in tods:
    val = round(dic_tod[tod]/total_trips*100, 5)
    desc = f'TOD_{tod}_PercentShare'
    new = pd.DataFrame({c1:[desc], c2:[val]})
    df_out = pd.concat([df_out, new])
df_out

total_vmt = 0
for tod in tods:
    df_links[f'{tod}_vmt'] = (df_links[f'autovol{tod.upper()}'] + df_links[f'truck volume{tod.upper()}'] +\
                              df_links[f'large truck linke volume{tod.upper()}']) * df_links['distance']
    val = df_links[f'{tod}_vmt'].sum()
    total_vmt += val
    desc = f'TOD_{tod}_VMT'
    new = pd.DataFrame({c1:[desc], c2:[val]})
    df_out = pd.concat([df_out, new])

new = pd.DataFrame({c1:['Total_VMT'], c2:[total_vmt]})
df_out = pd.concat([df_out, new])
df_out

# for per in periods:
#     df_temp.append({'Period': per,
#                     'Value': df[per+'_vmt'].sum()})
    
    
# df_links['am_vmt'] = (df_links['autovolAM'] + df_links['truck volumeAM'] + df_links['large truck linke volumeAM'])* df_links['distance']
# df_links['pm_vmt'] = (df_links['autovolPM'] + df_links['truck volumePM'] + df_links['large truck linke volumePM'])* df_links['distance']
# df_links['ea_vmt'] = (df_links['autovolEA'] + df_links['truck volumeEA'] + df_links['large truck linke volumeEA'])* df_links['distance']
# df_links['md_vmt'] = (df_links['autovolMD'] + df_links['truck volumeMD'] + df_links['large truck linke volumeMD'])* df_links['distance']
# df_links['ev_vmt'] = (df_links['autovolEV'] + df_links['truck volumeEV'] + df_links['large truck linke volumeEV'])* df_links['distance']

#### Write Out the Summary File

df_out.to_excel(out_path, sheet_name='Trip Summaries', index=False)
